+++ src/Credfeto.Package.Push/Constants/ExitCodes.cs
@@ -1,6 +1,6 @@
 namespace Credfeto.Package.Push.Constants;
 
-public static class ExitCodes
+internal static class ExitCodes
 {
     public const int Success = 0;
     public const int Error = 1;