+++ src/Credfeto.Package.Push/Searcher.cs
@@ -2,6 +2,7 @@ using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using Credfeto.Package.Push.Constants;
+using Credfeto.Package.Push.Helpers;
 
 namespace Credfeto.Package.Push;